Further additions

+++ Scripts/Xcorr/Run_Pack_Correlate.sh
@@ -33,10 +33,8 @@ stackFlag=1
 meanFlag=0
 sdFlag=0
 
-years=( `seq -w  2013 2017` )
+years=( `seq -w  2014 2017` )
 months=( `seq -w  1 12` )
-
-years=( '2015' )
 
 for year in "${years[@]}"; do
     for month in "${months[@]}"; do
@@ -51,36 +49,41 @@ for year in "${years[@]}"; do
 
 	#Determine the end day value within each month
 	echo ${year}/${month}
-	cd ${year}/${month}
-	find . -empty -type d -delete #deletes empty directories
-	endday=`ls -d [0-9][0-9] -v | tail -1`
-	cd ../../
 
-	echo $endday
+	#Check if associated year-month combination exists - is should, but if not bad things happen because of the
+	#cd ../../ command - change this to $basedir
+
+	if [ -d ${year}/${month} ]; then 
+	   cd ${year}/${month}
+	   find . -empty -type d -delete #deletes empty directories
+	   endday=`ls -d [0-9][0-9] -v | tail -1`
+	   cd $basedir
+
+	   echo $endday
 
 
         #./run_pack_spectrums.tcsh $year $month $startday $endday $StationList $basedir $db $deleteFlag $npts
 	#./run_cor_rp_single_spect.tcsh $db $outputCorDir $lag $stackFlag $meanFlag $sdFlag $basedir
 
-	echo $db
-	echo $outputCorDir
-	echo $lag
-	echo $stackFlag
-	echo $meanFlag
-	echo $sdFlag
-	echo $basedir 
+	   echo $db
+	   echo $outputCorDir
+	   echo $lag
+	   echo $stackFlag
+  	   echo $meanFlag
+	   echo $sdFlag
+	   echo $basedir 
 
-	./Pack_Spectrums.sh $year $month $startday $endday $StationList $basedir $db $deleteFlag $npts
+	   ./Pack_Spectrums.sh $year $month $startday $endday $StationList $basedir $db $deleteFlag $npts
 
-	echo $db
-	echo $outputCorDir
-	echo $lag
-	echo $stackFlag
-	echo $meanFlag
-	echo $sdFlag
-	echo $basedir
-	./Run_Correlation_Single_Spect.sh $db $outputCorDir $lag $stackFlag $meanFlag $sdFlag $basedir
-
+	   echo $db
+	   echo $outputCorDir
+	   echo $lag
+	   echo $stackFlag
+	   echo $meanFlag
+	   echo $sdFlag
+	   echo $basedir
+	   ./Run_Correlation_Single_Spect.sh $db $outputCorDir $lag $stackFlag $meanFlag $sdFlag $basedir
+       fi
     done
 done
